requirements.txt
# Core dependencies
streamlit>=1.32.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
plotly>=5.18.0
python-dotenv>=1.0.0
openai>=1.12.0
xlsxwriter>=3.1.9
python-dateutil>=2.8.2
python-redmine>=2.3.0
streamlit-option-menu>=0.3.2

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Development tools
black>=24.2.0
flake8>=7.0.0
mypy>=1.8.0

# Error handling and logging
python-json-logger>=2.0.7